refactor: move builder to a better place
