+++ src/query/service/tests/it/sql/planner/mod.rs
@@ -12,6 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod builders;
 mod format;
-mod select_builder;
 mod semantic;
